@@ -9,50 +9,52 @@
     "watch": "tsc -w",
     "test": "jest",
     "cdk": "cdk",
-    "prettier": "prettier -w .",
     "cdk-synth-env": "cdk synth env-aws-sso-extensions-for-enterprise -c config=env",
     "cdk-deploy-env": "cdk deploy env-aws-sso-extensions-for-enterprise -c config=env",
     "cdk-destroy-env": "cdk destroy env-aws-sso-extensions-for-enterprise -c config=env"
   },
   "devDependencies": {
-    "@aws-cdk/assert": "1.128.0",
+    "@aws-cdk/assert": "1.132.0",
     "@types/aws-lambda": "^8.10.85",
     "@types/jest": "^27.0.2",
-    "@types/node": "16.11.1",
-    "aws-cdk": "1.128.0",
-    "esbuild": "^0.13.12",
-    "eslint": "^8.0.1",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.25.2",
+    "@types/js-yaml": "^4.0.4",
+    "@types/node": "16.11.7",
+    "@typescript-eslint/eslint-plugin": "^5.3.1",
+    "@typescript-eslint/parser": "^5.3.1",
+    "aws-cdk": "1.132.0",
+    "esbuild": "^0.13.13",
+    "eslint": "^8.2.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-security": "^1.4.0",
     "jest": "^27.3.1",
     "prettier": "^2.3.2",
     "ts-jest": "^27.0.7",
-    "ts-node": "^10.3.0",
+    "ts-node": "^10.4.0",
     "typescript": "~4.4.4"
   },
   "dependencies": {
-    "@aws-cdk/aws-amplify": "1.128.0",
-    "@aws-cdk/aws-apigateway": "1.128.0",
-    "@aws-cdk/aws-dynamodb": "1.128.0",
-    "@aws-cdk/aws-events": "1.128.0",
-    "@aws-cdk/aws-events-targets": "1.128.0",
-    "@aws-cdk/aws-iam": "1.128.0",
-    "@aws-cdk/aws-kms": "1.128.0",
-    "@aws-cdk/aws-lambda": "1.128.0",
-    "@aws-cdk/aws-lambda-event-sources": "1.128.0",
-    "@aws-cdk/aws-lambda-nodejs": "1.128.0",
-    "@aws-cdk/aws-logs": "1.128.0",
-    "@aws-cdk/aws-s3": "1.128.0",
-    "@aws-cdk/aws-s3-notifications": "1.128.0",
-    "@aws-cdk/aws-sns": "1.128.0",
-    "@aws-cdk/aws-sns-subscriptions": "1.128.0",
-    "@aws-cdk/aws-sqs": "1.128.0",
-    "@aws-cdk/aws-ssm": "1.128.0",
-    "@aws-cdk/aws-stepfunctions": "1.128.0",
-    "@aws-cdk/core": "1.128.0",
-    "@aws-cdk/custom-resources": "1.128.0",
-    "@aws-cdk/pipelines": "1.128.0",
-    "cdk-watchful": "^0.5.269",
+    "@aws-cdk/aws-amplify": "1.132.0",
+    "@aws-cdk/aws-apigateway": "1.132.0",
+    "@aws-cdk/aws-dynamodb": "1.132.0",
+    "@aws-cdk/aws-events": "1.132.0",
+    "@aws-cdk/aws-events-targets": "1.132.0",
+    "@aws-cdk/aws-iam": "1.132.0",
+    "@aws-cdk/aws-kms": "1.132.0",
+    "@aws-cdk/aws-lambda": "1.132.0",
+    "@aws-cdk/aws-lambda-event-sources": "1.132.0",
+    "@aws-cdk/aws-lambda-nodejs": "1.132.0",
+    "@aws-cdk/aws-logs": "1.132.0",
+    "@aws-cdk/aws-s3": "1.132.0",
+    "@aws-cdk/aws-s3-notifications": "1.132.0",
+    "@aws-cdk/aws-sns": "1.132.0",
+    "@aws-cdk/aws-sns-subscriptions": "1.132.0",
+    "@aws-cdk/aws-sqs": "1.132.0",
+    "@aws-cdk/aws-ssm": "1.132.0",
+    "@aws-cdk/aws-stepfunctions": "1.132.0",
+    "@aws-cdk/core": "1.132.0",
+    "@aws-cdk/custom-resources": "1.132.0",
+    "@aws-cdk/pipelines": "1.132.0",
+    "cdk-watchful": "^0.5.292",
     "js-yaml": "^4.1.0",
     "source-map-support": "^0.5.16"
   }
