@@ -1,21 +1,27 @@
 module.exports = {
   env: {
-    browser: true,
-    commonjs: true,
-    es2021: true,
+    node: true,
   },
-  extends: ['airbnb-base'],
+  ignorePatterns: ["**/*.js", "cdk.out"],
+  extends: [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:security/recommended",
+  ],
+  parser: "@typescript-eslint/parser",
   parserOptions: {
     ecmaVersion: 12,
+    sourceType: "module",
   },
+  plugins: ["@typescript-eslint", "security"],
   rules: {
-    'no-console': 'off',
-    'import/no-unresolved': 'off', // needed for ignoring dependencies that we manage through lambda layers
-    'no-useless-escape': 'off',
-    'no-restricted-syntax': 'off',
-    'no-await-in-loop': 'off',
-    'import/no-extraneous-dependencies': 'off', // needed for ignoring dependencies that we manage through lambda layers
-    'func-names': 'off', // to override the async await generators being generated for pagination
-    eqeqeq: ['error', 'smart'],
+    "linebreak-style": ["error", "unix"],
+    "@typescript-eslint/no-non-null-assertion": "off",
+    eqeqeq: ["error", "always", { null: "ignore" }],
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-unused-vars": [
+      "warn",
+      { argsIgnorePattern: "^_", varsIgnorePattern: "^_" },
+    ],
   },
 };
