@@ -12,8 +12,8 @@ Trigger source: permission set DDB table stream notifications for NEW_AND_OLD_IM
 const { topicArn, errorNotificationsTopicArn, AWS_REGION } = process.env;
 
 // SDK and third party client imports
-const { SNSClient, PublishCommand } = require('@aws-sdk/client-sns');
-const DynamoStreamDiff = require('dynamo-stream-diff');
+const { SNSClient, PublishCommand } = require("@aws-sdk/client-sns");
+const DynamoStreamDiff = require("dynamo-stream-diff");
 
 // SDK and third party client object initialistaion
 const snsClientObject = new SNSClient({ region: AWS_REGION });
@@ -23,7 +23,8 @@ const errorParams = {
   TopicArn: errorNotificationsTopicArn,
 };
 const errorMessage = {};
-errorMessage.Subject = 'Error Processing Permission set DDB table stream records';
+errorMessage.Subject =
+  "Error Processing Permission set DDB table stream records";
 
 exports.handler = async (event) => {
   try {
@@ -35,28 +36,28 @@ exports.handler = async (event) => {
         const message = {};
         message.event_type = record.eventName;
         message.record = record;
-        if (record.eventName === 'INSERT') {
+        if (record.eventName === "INSERT") {
           // insert recieved
-        } else if (record.eventName === 'MODIFY') {
+        } else if (record.eventName === "MODIFY") {
           const diff = DynamoStreamDiff(record);
           message.diff = diff;
-        } else if (record.eventName === 'REMOVE') {
+        } else if (record.eventName === "REMOVE") {
           // remove record
         }
         params.Message = JSON.stringify(message);
         await snsClientObject.send(new PublishCommand(params));
         console.log(
           `Successfully posted to permissionset topic from permissionset stream handler content: ${JSON.stringify(
-            message,
-          )}`,
+            message
+          )}`
         );
-      }),
+      })
     );
-    return 'Successfully processed all records in the current stream shard';
+    return "Successfully processed all records in the current stream shard";
   } catch (err) {
     console.error(
-      'main handler exception in permissionSet stream handler: ',
-      err,
+      "main handler exception in permissionSet stream handler: ",
+      err
     );
     errorMessage.event_detail = event;
     errorMessage.error_details = err;
